SMS_tx: Remove clash between auto and manual message interval modification
Previously, air_sms is designed to either automatically increase its messaging frequency when detecting that data telem is down, or upon request ("SMS short") from Ground Control. However, both these features clash with each other and prevent the "SMS short" feature from working in Padi5

The easiest way to fix this bug is to remove the auto message frequency feature with data telem. It will still report the status of data telemetry, but the onus is now on the user to manually increase the messaging frequency.